ref: extract constants to separate folder
